Extract XML classes to individual files
